+++ uefi/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `proto::media::block::BlockIO2`.
 - Added support for `revision()` and `device_type_guid()` to `Serial` protocol
 - Added `Serial::read_exact()` and `Serial::write_exact()`
+- Added `Serial::read_to_vec()`
 
 ## Changed
 - Changed ordering of `proto::pci::PciIoAddress` to (bus -> dev -> fun -> reg -> ext_reg).